+++ database/migrations/2015_12_24_095809_create_stock_outs.php
@@ -17,7 +17,7 @@ class CreateStockOuts extends Migration
             $table->integer('item_id')->comment('item号')->default(NULL);
             $table->string('sku')->comment('sku')->default(NULL);
             $table->integer('amount')->comment('数量')->default(NULL);
-            $table->integer('total_amount')->comment('总金额')->default(NULL);
+            $table->float('total_amount')->comment('总金额')->default(NULL);
             $table->integer('warehouses_id')->comment('仓库id')->default(NULL);
             $table->integer('warehouse_positions_id')->comment('库位id')->default(NULL);
             $table->string('type')->comment('出库类型')->default(NULL);
